Fix remarks
