+++ addon/objects/uml-primitives/fd-uml-active-state.js
@@ -3,8 +3,7 @@
 */
 
 import Ember from 'ember';
-
-import { BaseObject } from './fd-uml-baseobject';
+import { State } from './fd-uml-state';
 import FdUmlObject from './fd-uml-baseobject';
 
 /**
@@ -24,58 +23,13 @@ export default FdUmlObject.extend({
   state: Ember.computed.alias('primitive.Text.Text'),
 
   /**
-    Type of primitive.
-
-    @property type
-    @type String
-  */
-  type: Ember.computed.alias('primitive.$type'),
-
-  /**
     See {{#crossLink "FdUmlPrimitive/JointJS:method"}}here{{/crossLink}}.
 
     @method JointJS
   */
   JointJS() {
     let properties = this.getProperties('id', 'size', 'name', 'position', 'state');
-    if (this.get('type') === 'STORMCASE.UML.std.State, UMLSTD') {
-      return new State(properties);
-    } else if (this.get('type') === 'STORMCASE.UML.std.Class, UMLSTD') {
-      return new StdClass(properties);
-    } else {
-      return new ActiveState(properties);
-    }
-
-  },
-});
-
-/**
-  Defines the JointJS object, which represents a 'StdClass' object in the UML diagram.
-
-  @for FdUmlActiveState
-  @class StdClass
-  @extends BaseObject
-  @namespace flexberry.uml
-  @constructor
-*/
-export let StdClass = BaseObject.define('flexberry.uml.stdClass', {
-  attrs: {
-    'text': { 'font-weight': 'bold' }
-  }
-});
-
-/**
-  Defines the JointJS object, which represents a 'State' object in the UML diagram.
-
-  @for FdUmlActiveState
-  @class State
-  @extends StdClass
-  @namespace flexberry.uml
-  @constructor
-*/
-export let State = StdClass.define('flexberry.uml.State', {
-  attrs: {
-    '.flexberry-uml-header-rect': { rx:10, ry:10 },
+    return new ActiveState(properties);
   }
 });
